Replacing bigtable-protos-internal jar with generated java.
The bigtable-protos project was causing a problem with the release strategy.  The release plugin doesn't allow -SNAPSHOT dependencies, and the protos jar was a -SNAPSHOT version.  Instead of just checking in a non SNAPSHOT version, I think that it's better to check in java files generated from gradle's generateProto process.  That simplified the pom.xml and seems more transparent.

Moving bigtable-protos/src/main/generated-java to src/generated/java