Add support for HBase PageFilters.
Add a PageFilterAdapter which will adapt PageFilter instances so long as
they appear at the top-level or as the last element of the top-most
FilterList.

Add the concept of ReadHooks. Right now, ReadHooks are for modifying a
ReadRowsRequest before sending the request to CBT.

Add a ReadOperationAdapter interface for Scan and Get adapters and move
from plain OperationAdapter to ReadOperationAdapter.