Allowing a user to retry even if the client does not set the timestamp. (#800)
The concern about setting the client timestamp in hbase is safe from the standpoint of most situations.  A Column Family with a single version and gets/scans where version=1 will not have a problem with retries that put in two cells into a qualifier.  Allow a user to set an option that will ensure that a server-side timestamp is set.