More javadoc fixes (#927)
* Exclude v1 API from javadoc generation

I think it's best to skip this set since the V1 API is deprecated (and
it generates a bunch of warnings). We can always change our minds later
of course if we have to revisit this.

Testing Done:
- verified that the warnings produced during javadoc:javadoc for v1
  protos go away

* Remove executable permissions from README.md

No need for +x on this markdown file.

* Autofix a large number of javadoc warnings

This change is the result of using javadoc:fix to mostly automatically
fix a huge set of javadoc warnings.

There were a couple of cases where I had to fix up what the autofixer
produced:

BigtableOptions.java
- for some reason the autofixer inserted copies of a large number of
  functions and members resulting in duplicated / redefined names and a
  non-compiling file. Fixed it manually after inspecting the damage.

BulkOptions.java
- tool didn't know to change '<' to '&lt;' and so generated bad HTML.

author tags:
The tool inserted @author dharan all over the place and I manually
changed it to sduskis since that seemed like the logical default.

Testing Done:
- mvn javadoc:javadoc (still produces warnings, but less of them now)
- mvn clean install (all tests pass)

* Silence the javadoc linter

The javadoc linter in Java 8 produces a larger number of warnings that
are not easily fixable for us since they depend on upstream fixes, see
e.g. https://github.com/google/protobuf/issues/304

So just run without the linter for now.

* Fix remaining javadoc warnings

- fix a couple of warnings for @link cases that javadoc was unable to
  resolve, and add a link to the gRPC docs and make sure
  io.grpc.Status.Code can be resolved
- Remove autofixer's mistakenly applied @inheritDoc
- fixup documentation for Logger class to better explain varargs use