Fix javadoc:aggregate-jar target
This change makes the javadoc:aggregate-jar target work at the
top-level.

Doing this correclty is complicated because javadoc:aggregate is
confused by the repackaged/shadedPattern renaming of the bigtable client
in the dataflow submodules and fails complaining about
missing/unresolveable symbols/packages.

After a lot of experimentation, banging my head against the wall, and
discussions with Jeff I decided the best way to handle javadoc at the
top-level was simply to exclude the dataflow projects; we already have a
way to handle and publish those docs that we aren't going to change in
the near term, so we simply skip the com.google.cloud.bigtable.dataflow*
packages when doing a top-level javadoc:aggregate[-jar] for the main
project.

I also made a change to ensure that you can turn back on the doclint
behavior by passing -Ddoclint on the commandline if you want to inspect
it for any javadoc target, though it's still not particularly useful
given all the noise from grpc-java (and note that they have it disabled
in their build).

I also made some tweaks to the way we express configuration so the
javadoc stuff will apply for all profiles, and the aggregate-jar target
will be used during the package phase.

Finally I threw in a couple of fixes for new warnings that I noticed.