Added Stack Overflow tag badge; removed ghit.
The functionality that ghit provides (visit counts) can be easily viewed by
project owners via the console, but in itself is not as useful to visitors as
other badges such as build status, Maven central, or the Stack Overflow tag.

Also moved all references to badges and links to the bottom of the page for
readability.