Adding adapters for HBase Mutations -> MutateRowsRequest.Entry. (#1112)
BigtableBufferedMutor and BatchExecutor should convert from a Put/Delete directly to a MutateRowsRequest.Entry rather than first converting to a MutateRowRequest and then to a MutateRowsRequest.Entry. This PR adds the ability to convert to a MutateRowsRequest.Entry. A future PR will use the new ability from BufferedMutator and BatchExecutor.