Performance increase for scanning. Results from Bigtable are already ordered, although not ideally for HBase.  The order we get: - family - some internally ordered mechanism for qualifier - reverse timestamp within the qualifier
Currently, we don't assume any ordering when we convert from the results we get from Cloud Bigtable.  We currently dedup using a TreeMap orderd by family/qualifier/timstamp via an HBase Comparator.  That's an expensive operation, or at least more expensive than it could be.  With the the assumptions of the server returning the order above, we can dedup within a single qualifer by simply checking against the previously returned response to see if the key matches, and if so, omit the value.  We can sort qualifier/timestamp within a family rather than sorting all cells across all families.

These changes made a 2-3X difference in the performance of the FlatRowAdapter as measured by RowMergerPerf.