BigtableClusterUtilities should have no-arg methods. (#1124)
* BigtableClusterUtilities should have no-arg methods.
BigtableClusterUtilities allows users to programmatically control the size of their cluster.  It is generally used with a user configuring a project and an instance.  A user can also set '-' as the instance id for all projects, which is useful when we know the cluster id but not the instance id.

Currently, one project/instance combination which only have a single cluster and zone is likely to be our most common scenario.  As such, users should not have to specify clusterId and zoneId.  In those scenarios, the code should look like:

`BigtableClusterUtilities.forInstance(projectId, instanceId).setClusterSize(30);`

It currently looks like:

`BigtableClusterUtilities.forInstance(projectId, instanceId).setClusterSize(clusterId, zone, 30);`

Add new methods for setSize, incrementSize, getClusterSize that don't have clusterId and zone.

* Update BigtableClusterName.java

* Addressing some comments.
- BigtableClusterName.getClusterName() becmoes getClusterId()
- Removing BigtableClusterUtilities.incrementXX methods.
- Updating comments

* Updating comments.