Fix ordering of ScanAdapter filter chaining (#1229)
Max versions implies get the top N remaining cells.  Order matters with chaining, and the order was previously broken.  In most cases where a Column Family is configured with a single latest cell this doesn't matter.  In cases where there are multiple cells, and the user wants an earlier cell, the ordering of the chain of filters has to change from

1) max versions filter
2) timestamp filter
3) other filters

to

1) timestamp filter
2) other filters
3) max version filter