Remove paths to HConstants (#1254)
* Remove paths to HConstants
Some confluence of events in Dataflow caused a deadlock.  There were some strange stack traces that pointed towards the root cause being the class loading of HConstants.  Scans refer to HConstants.  KeyValues refer to HConstants.  Puts and Adapters -> KeyOnlyFilterAdapter both refer to HContants.

This PR removes a couple of paths to HConstants in hope that this removes the class loading issue.

* Preloading HConstants.class to avoid deadlock
KeyValue and HConstant depend on each other, which can cause deadlocks.

* Reverting KeyOnlyFilterAdapter.java and using Adapters.class.getName().