Tweaking thread pool creations: Use GrpcUtil. (#1266)
The grpc library works well on AppEngine Java 7, so this PR leverages GrpcUtil as a fix for #1243.  While doing the investigation of GrpcUtil, I found that a lot of the work we did 2 years ago to use daemon threads was worked into the gRPC library, so we can get rid of our custom configuration in gRPC for "elg" and the threadpool.