Adding a MutateRowsRequestManager (#1570)
* Adding a MutateRowsRequestManager
This new class encapsulates the logic of handling the MutateRowsRequest/Response interaction across retries.  This change is very similar to the structure of ReadRowsRequestManager / RetryingReadRowsOperation.

* Adding an enum instead of multiple booleans.

* Fixing typos.  Also adding volatile where necessary.
I reviewed this with @igorbernstein, and the volatile changes were deemed necessary for any case where an object would be used across threads.  If this is a problem for MutateRows, then it would be a problem for ReadRows as well, so the ReadRows operation was also updated.

* Improving testing.

* Fixing a signficant issue with RetryingReadRowsOperation, and shoring up TestMutateRowsRequestManager.
Fixing a minor issue with RetryingReadRowsOperationTest