SingleColumnValueFilterAdapter now uses an interleave. (#1604)
* SingleColumnValueFilterAdapter now uses an interleave.
`SingleColumnValueFilterAdapter` is a relatively complex HBase filter.  There were circumstances where it creates nested `Condition`s, which is problematic on the server-side.  A user used multiple `SingleColumnValueFilterAdapter`s which in turn caused severe problems.  @dmmcerlean suggested an alternative that would use `Interleave`d `Condition`s instead of nested `Condition`s.  This PR implements the suggestion.  In addition to the implementation itself, this PR attempts to document the intent of `SingleColumnValueFilterAdapter` as well as how to convert the options therein to a Cloud Bigtable `RowFilter`.

In addition, `TestSingleColumnValueFilter` was updated to test new conditions, and all `Scan`s were constrained to only the range of rows created by the test.

Also, TestSnapshots used keys that contained `TestSingleColumnValueFilter`.  The keys were updated to use a prefix of `TestSnapshots` instead.

* Updating the comments.