Removing table truncation in AbstractTest (#1620)
The HBase minicluster test fails in a weird way.  Disabling the table times out at the end of the test.  The current tests don't require truncation.  Truncation is a future proofing mechanism which is slowing down tests and causing HBase minicluster to fail.  We can remove the truncation now, and figure out better alternatives in the future... YAGNI.