Decoupling AbstractTest and IntegrationTests. (#1639)
`IntegrationTets` needs to set up the shared test environment, and `AbstractTest` and its subclasses need to use that environment.  This PR adds a level of indirection for the shared test environment by adding a singleton for the shared test environment.  IntegrationTests will set up the singleton, and `AbstractTest` will get that singleton.

The goal is to keep `IntegrationTests` where they are, and move `AbstractTest` and common subclasses to bigtable-hbase-integration-tests-common.