Fixing a broken filter test (#1652)
Our hbase tests are failing on TestFilters.testMultiRangeFilter().  There is a potential conflict between testMultiRangeFilterOrList() and testMultiRangeFilter() in terms of key space, which may be the root of the problem.  I could not reporoduce the problem myself, but as part of my investigation I found that testMultiRangeFilterOrList() no longer needs @Category(KnownGap.class).