Fixing flaky OauthCredentialsCache test. (#1754)
There was a race condition in OAuthCredentialsCacheTest.testRefreshAfterStale().  The test is for a stale condition that should ensure that the token's cache state is `CacheState.Stale`.  The test would be consistent if only a single call is made to the core cache retrieval code; however, a second call to `getHeader()` was introduced and that second call may return the "Stale" token or the "Good" token depending on the execution of a different thread.
This change ensures that only a single call is made to refresh the header cache under stale conditions.