RefreshingOAuth2CredentialsInterceptor TIMEOUT is now final (#1780)
The TIMEOUT existed as a way for users to override the default waiting behavior.  TIMEOUT was not final, which made it problematic as per issue #1775.  After changing TIMEOUT to 15 seconds (PR #1657), and using CallOptions deadline (#1727), users ought to have a much better experience with credentials refresh.
This PR removes threetenbp and replaces it with a long and an explicit TimeUnit.