Fix a couple of Throttler issues:
The common path to invoking the ThrottlingClientInterceptor:
call = channel.newCall(...)
call.sendMessage(request)
call.halfClose()

However if the resource limiter is interrupted while the interceptor is
waiting to send a message, the delegate call will not be set, causing an
NPE.

To fix this issue, the ForwardedClientCall was remove and all forwarded
calls check for null delegates. The impl for halfClose is skipped
entirely because only UnaryCalls are supported.

The other issue addressed here is an unhandled cancellation of the call
before sending a message.