Adding AutoCloseable to IBulkMutation
It seems existing [BulkMutation]() does not have close() so I extended `AutoCloseable` to `IBulkMutation`.

Currently BigtableBufferedMutator closes the resources in it's close()  so I marked `IBulkMutation` closed in the same.

This will help us to close the GCJ's `BulkMutationBatcher` instance in the current BulkMutationGCJClient, and in next PRs we can close the new `Batcher` instance as well.