chore: adding solution for flaky test case
Problem:
There is a race condition between cleanUp() and batcher's callback. The cleanUp() calls -> `batcher.close()` -> `flush()` -> `awaitAllOutstandingBatches()` and the `BatcherImpl#onBatchComplemetion` callback decrements `numOfOutstandingBatches`, It notifies all once it reaches **0**.
This led to a situation where `BatcherImpl#awaitAllOutstandingBatches()` grabs the `flushLock` and wait for notify & The `BatcherImpl#onBatchCompletion` never gets a chance to decrement the counter and notify.

This only happened until now in case we called `BulkReadVeneerApi#sendOutstanding()`. This method decrements the `cleanupBarrier` on each `sendOutstanding` call, which led to counter being **0** before all the entries could actually be resolved.

Solution:
With this commit, I moved the counter from `notifyArrival()` to `ApiFuture<Result>`'s callback, and added a `signalCleanUp` flag to trigger batcher cleanups.

The `newSingleThreadExecutor` is also needed to handle a case where the user calls `BulkReadVeneerApi#sendOutstanding` after adding only one `rowKey` in batch, This creates the above scenario of waiting on `flushLock`.