fix: race condition where a retry triggers a full table scan (#2758)
* fix: race condition where a retry triggers a full table scan (#2742)

* fix: fix race condition where a retry triggers a full table scan

There is a race condition where a read rows stream will return all of the data, but still end in a non-ok status. This will trigger a full table scan because the row resumption logic will strip away all of seem keys & ranges, leaving the retry request empty, will cause a full table scan.

This PR will detect this case and will replace the status code to be ok. This should be a very rare condition.

* fix broken test

* fix broken test2

* format

* fix broken test