fix: hbase 2x shell (#2905)
fixes #2866
bigtable hbase shell has been broken since hbase 2.2.0. It broke because hbase added a feature that the shell referenced. The feature is not actively used by the shell, but it is referenced. This PR defers the unsupported operation errors so that the shell can load

Creating an integration test is going to take a bit more work, so it was extracted to #2904