chore(deps): shade but dont relocate bigtable-metrics-api
bigtable-metrics-api contains a public surface to configure custom metrics. So the surface can't be mangled.
However, it depends on api-common, which pulls in heavy deps like guava. And since bigtable-metrics-api needs to be exposed by
bigtable-hbase, we need to expose its api w/o pulling in cconflicting versions of guava