feat: improve handling when crossing async and sync boundaries
Previously we would blindly uses ApiExceptions.callAndTranslateApiException and Futures.getUnchecked to await future results. This created a number of problems:
* It would bubble ApiExceptions & CheckExceptions to the surface, which is quite unexpected for an hbase client
* It would hide hbase specific IOExceptions

This PR introduces FutureUtil#unwrap, which will try to preserve the original exception. However to aid in debugging, it will augment the stacktrace to include the caller. This is a similar approach as hbase took in:
https://github.com/apache/hbase/blob/5b9940907e695e20d24968cbc725277a19ce2170/hbase-common/src/main/java/org/apache/hadoop/hbase/util/FutureUtils.java#L129-L140

Unfortunately we can't just call that utility because it only exists in hbase 2x, so I inlined relevant bits