chore: make bigtable-hbase a drop in migration tool part 1 - hbase 1 libraries (#4014)
* chore: make bigtable-hbase a drop in migration tool part 1 - hbase 1 libraries

This should make bigtable-hbase's role a bit clearer: to be a drop in shim to help migrate from hbase to bigtable for existing applications instead of being a standalone client. To this end, all library jars will weaken their dependency on hbase-client to be `provided` by the enduser. The primary interaction will be via the -shaded jars tgat endusers will be able to drop into their $HBASE_CLASSPATH and then connect to bigtable using a tweaked hbase-site.xml. This interaction will emphasis minimal code changes.

This PR focuses on HBase 1 libraries:
- all dependencies on hbase are changed to provided scope
- `-hadoop` modules are removed since they are identical to `-shaded` modules
- maven plugin to ensure that we don't clobbered provided dependencies
- shading configs simplified due to provided scope
- removal of slf4j version enforcement since those deps are fully provided by hbase now

Change-Id: I7946c06284c81caa0a28fc7ae4773d85fab4381a

* tweaks

Change-Id: I085b4e1371a008c23a59a5f69673eb0b08ef7471

* format

Change-Id: I32d3cda81941927cc14d3609cd73db30a2ef5682