deps: add a direct dep on google-cloud-bigtable in the -shaded artifacts to workaround UpperBound issues (#4244)
Both google-cloud-bigtable & opencensus exporters have a transitive deps on checker-qual. checker-qual is only used for annotations and doesnt actually affect anything in the shaded jars. In fact it is marked as provided. It doesnt matter what version (if any) is present at runtime. However RequireUppperBounds enforcer check trips over this. So as a workaround, this PR adds a direct dep on google-cloud-bigtable in the -shaded artifacts. This should have no effect on the final build as that artifact will be shaded regardless

Change-Id: I44a000379d822c28daf4dc27a41d1d8f6998689f