test: fix dependency conflict in integration tests
bigtable-hbase-*-integration-tests declare a dependency on truth assertion library. truth has a transitive dependency on guava-*-android. Because of maven's dependency resolution, this dependency overrides the guava version that bigtable-hbase & java-bigtable need which is the jre version of guava. To mitigate the conflict, this change excludes truth's transitive dep on guava

Change-Id: Iea16d837f71d2b4f8f037fa870fa871dcc2fb64b