fix: align the handling of regular expressions between hbase & bigtable
When using RegexStringComparator, HBase will try to match a substring of the target. While bigtable requires a full match. For example, given the target "abcd", in HBase the regex "[bc]*" will match the string. However Bigtable will not. Bigtable will require the regex to also match a & d.

This PR addresses this discrepancy by surrounding the user provided regex with a zero or more wildcards. ie transform "[bc]*"  into ".*[bc]*.*"

Change-Id: I7404bddd4c9c9d907d57225d99f7062c4407fc35