fix: align the handling of regular expressions between hbase & bigtable (#4430)
* fix: align the handling of regular expressions between hbase & bigtable

When using RegexStringComparator, HBase will try to match a substring of the target. While bigtable requires a full match. For example, given the target "abcd", in HBase the regex "[bc]*" will match the string. However Bigtable will not. Bigtable will require the regex to also match a & d.

This PR addresses this discrepancy by surrounding the user provided regex with a zero or more wildcards. ie transform "[bc]*"  into ".*[bc]*.*"

Change-Id: I7404bddd4c9c9d907d57225d99f7062c4407fc35

* avoid wrapping empty strings in wildcards

Change-Id: I60325d415f24fe6dbd02fa84978047e5759128cf

* update unit tests

Change-Id: I617e968dc0d2600952f2e06d1789b6e9e9539e6b

* remove stray debug statement

Change-Id: Ie07e3c5d4beb03af93af8dbfa98847010f90457d

* fix typo in regex test and re-order the test cases to make it easier to catch the typo

Change-Id: I7bdbf601984e32f0f4b3703e19f5f61fc476967d